Rename SvgBuilder into WithSvg.
