+++ examples/walk_path/src/main.rs
@@ -38,7 +38,7 @@ fn main() {
     let tolerance = 0.002;
 
     // Build a Path for the rust logo.
-    let mut builder = Path::svg_builder();
+    let mut builder = Path::builder().with_svg();
     build_logo_path(&mut builder);
     let logo_path = builder.build();
